backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
yfinance==0.2.32
pydantic==2.5.0
python-dotenv==1.0.0
aiohttp==3.9.1
pytest==7.4.3
httpx==0.25.2
